Second Version for Testing
Here I made the end signal of the bluetooth app an undoable amount of
steps to verify the program is done. Once the stepper motor task has
caught the "0" number of steps. The finishing process begins where we
verify the use has taken off the syringe and we can start reversing.